// PC.BE/src/common/enums/shortcut.enums.ts
export enum TypeShortcut {
  shortcutTabs = 'shortcut::tabs',
  shortcutCards = 'shortcut::cards',
}

export enum StyleCardShortcut {
  cardsShortcutSmall = 'cards:shortcut::small',
  cardsShortcutMedium = 'cards:shortcut::medium',
  cardsShortcutLarge = 'cards:shortcut::large',
}

export enum StyleAlignShortcut {
  alignLeft = 'align::left',
  alignCenter = 'align::center',
  alignRight = 'align::right',
}

export enum TargetShortcut {
  targetHome = 'target::home',
  targetBusiness = 'target::business',
  targetEnterprise = 'target::enterprise',
  targetInsurance = 'target::insurance',
}